Windows multiprocessing compatibility.
The lack of certain feature in Windows (os.for()) means that certain bugs not
noticeable in liux become apparent. Namely, child process spawn by making use of
multiprocessing pass serialised data between them. Python uses pickle to do this,
and pickle cannot serialise bound or unbound methods. Therefore, the arguments of
these processes must be top level functions.

I accomplished this by putting the two core functions of the class in charge of
the connections at the top level, and defining the queue and subprocess inside the
main class.